+++ test.cpp
@@ -1,0 +1,32 @@
+#include <iostream>
+#include "LinearAlgebra/gauss_seidel.h"
+
+using namespace std;
+
+int main() {
+    const int n = 4;
+
+    double** A = new double*[n];
+
+    for (int i = 0; i < n; ++i) {
+        A[i] = new double[n];
+    }
+
+    // Define a matrix (4D array)
+    A[0][0] = 10.0; A[0][1] = -1.0; A[0][2] = 2.0; A[0][3] = 0;
+    A[1][0] = -1.0; A[1][1] = 11.0; A[1][2] = -1.0; A[1][3] = 3.0;
+    A[2][0] = 2.0; A[2][1] = -1.0; A[2][2] = 10.0; A[2][3] = -1.0;
+    A[3][0] = 0.0; A[3][1] = 3.0; A[3][2] = -1.0; A[3][3] = 8.0;
+    double b[n] = {6.0, 25.0, -11.0, 15.0};
+    double x[n] = {0.0, 0.0, 0.0, 0.0};
+
+    GaussSeidelMethod(A, b, x, n);
+
+    // Clean up dynamic memory
+    for (int i = 0; i < n; ++i) {
+        delete[] A[i];
+    }
+    delete[] A;
+
+    return 0;
+}
